pyproject.toml
[tool.poetry]
name = "pet-store"
version = "0.1.0"
description = "pet store auction REST API"
authors = ["Adham Ahmed <adham2272001@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.3"
djangorestframework = "^3.15.1"
djoser = "^2.2.2"
djangorestframework-simplejwt = "^5.3.1"
drf-nested-routers = "^0.93.5"
drf-yasg = "^1.21.7"


[tool.poetry.group.dev.dependencies]
pytest = "^8.1.1"
pytest-django = "^4.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
